+++ src/analyzer.c
@@ -142,7 +142,7 @@ void* analyzer_thread_function(void* args) {
     pthread_exit(&thread->return_status);
 }
 
-AnalyzerThread* analyzer_thread_create(LoggerThread* logger,ReaderData* reader_data, unsigned short thread_count, size_t buffer_length) {
+AnalyzerThread* analyzer_thread_create(LoggerThread* logger,ReaderData* reader_data, unsigned int thread_count, size_t buffer_length) {
     AnalyzerThread* analyzer_thread = malloc(sizeof(AnalyzerThread));
     analyzer_thread->logger = logger;
     analyzer_thread->should_end = 0;
